switched DomainContext to CurrentDomainContext
In beta 13 a block performance change changed the API, which the
DomainContext depended on. CurrentDomainContext is added to comply with
the API change.
