+++ domain/domain.services.yml
@@ -5,10 +5,10 @@ services:
       - { name: access_check }
     arguments: ['@domain.negotiator']
   block.domain_context:
-    class: Drupal\domain\EventSubscriber\DomainContext
+    class: Drupal\domain\ContextProvider\CurrentDomainContext
     arguments: ['@domain.negotiator']
     tags:
-      - { name: 'event_subscriber' }
+      - { name: 'context_provider' }
   domain.creator:
     class: Drupal\domain\DomainCreator
     tags:
